bug修复

+++ YouFeel/Library/HistogramView/BXTHistogramStatisticsView.m
@@ -42,7 +42,7 @@
             secondkwhLabel.font = [UIFont systemFontOfSize:11.f];
             secondkwhLabel.adjustsFontSizeToFitWidth = YES;
             secondkwhLabel.textAlignment = NSTextAlignmentRight;
-            secondkwhLabel.text = [NSString stringWithFormat:@"%ld",measure/6*5];
+            secondkwhLabel.text = [NSString stringWithFormat:@"%ld",(long)measure/6*5];
             [self addSubview:secondkwhLabel];
             
             UILabel *thridkwhLabel = [[UILabel alloc] initWithFrame:CGRectMake(0.f, 120.f, 30.f, 20.f)];
@@ -50,7 +50,7 @@
             thridkwhLabel.font = [UIFont systemFontOfSize:11.f];
             thridkwhLabel.adjustsFontSizeToFitWidth = YES;
             thridkwhLabel.textAlignment = NSTextAlignmentRight;
-            thridkwhLabel.text = [NSString stringWithFormat:@"%ld",measure/6*4];
+            thridkwhLabel.text = [NSString stringWithFormat:@"%ld",(long)measure/6*4];
             [self addSubview:thridkwhLabel];
             
             UILabel *fourthkwhLabel = [[UILabel alloc] initWithFrame:CGRectMake(0.f, 170.f, 30.f, 20.f)];
@@ -58,7 +58,7 @@
             fourthkwhLabel.font = [UIFont systemFontOfSize:11.f];
             fourthkwhLabel.adjustsFontSizeToFitWidth = YES;
             fourthkwhLabel.textAlignment = NSTextAlignmentRight;
-            fourthkwhLabel.text = [NSString stringWithFormat:@"%ld",measure/2];
+            fourthkwhLabel.text = [NSString stringWithFormat:@"%ld",(long)measure/2];
             [self addSubview:fourthkwhLabel];
             
             UILabel *fifthkwhLabel = [[UILabel alloc] initWithFrame:CGRectMake(0.f, 220.f, 30.f, 20.f)];
@@ -66,7 +66,7 @@
             fifthkwhLabel.font = [UIFont systemFontOfSize:11.f];
             fifthkwhLabel.adjustsFontSizeToFitWidth = YES;
             fifthkwhLabel.textAlignment = NSTextAlignmentRight;
-            fifthkwhLabel.text = [NSString stringWithFormat:@"%ld",measure/3];
+            fifthkwhLabel.text = [NSString stringWithFormat:@"%ld",(long)measure/3];
             [self addSubview:fifthkwhLabel];
             
             UILabel *sixthkwhLabel = [[UILabel alloc] initWithFrame:CGRectMake(0.f, 270.f, 30.f, 20.f)];
@@ -74,7 +74,7 @@
             sixthkwhLabel.font = [UIFont systemFontOfSize:11.f];
             sixthkwhLabel.adjustsFontSizeToFitWidth = YES;
             sixthkwhLabel.textAlignment = NSTextAlignmentRight;
-            sixthkwhLabel.text = [NSString stringWithFormat:@"%ld",measure/6];
+            sixthkwhLabel.text = [NSString stringWithFormat:@"%ld",(long)measure/6];
             [self addSubview:sixthkwhLabel];
             
             UILabel *seventhkwhLabel = [[UILabel alloc] initWithFrame:CGRectMake(0.f, 320.f, 30.f, 20.f)];
@@ -163,7 +163,7 @@
             secondkwhLabel.font = [UIFont systemFontOfSize:11.f];
             secondkwhLabel.adjustsFontSizeToFitWidth = YES;
             secondkwhLabel.textAlignment = NSTextAlignmentRight;
-            secondkwhLabel.text = [NSString stringWithFormat:@"%ld",measure/4*3];
+            secondkwhLabel.text = [NSString stringWithFormat:@"%ld",(long)measure/4*3];
             [self addSubview:secondkwhLabel];
             
             UILabel *thridkwhLabel = [[UILabel alloc] initWithFrame:CGRectMake(0.f, 170.f, 35.f, 20.f)];
@@ -171,7 +171,7 @@
             thridkwhLabel.font = [UIFont systemFontOfSize:11.f];
             thridkwhLabel.adjustsFontSizeToFitWidth = YES;
             thridkwhLabel.textAlignment = NSTextAlignmentRight;
-            thridkwhLabel.text = [NSString stringWithFormat:@"%ld",measure/2];
+            thridkwhLabel.text = [NSString stringWithFormat:@"%ld",(long)measure/2];
             [self addSubview:thridkwhLabel];
             
             UILabel *fourthkwhLabel = [[UILabel alloc] initWithFrame:CGRectMake(0.f, 245.f, 35.f, 20.f)];
@@ -179,7 +179,7 @@
             fourthkwhLabel.font = [UIFont systemFontOfSize:11.f];
             fourthkwhLabel.adjustsFontSizeToFitWidth = YES;
             fourthkwhLabel.textAlignment = NSTextAlignmentRight;
-            fourthkwhLabel.text = [NSString stringWithFormat:@"%ld",measure/4];
+            fourthkwhLabel.text = [NSString stringWithFormat:@"%ld",(long)measure/4];
             [self addSubview:fourthkwhLabel];
             
             UILabel *fifthkwhLabel = [[UILabel alloc] initWithFrame:CGRectMake(0.f, 320.f, 35.f, 20.f)];
